+++ tests/hazmat/primitives/decrepit/test_algorithms.py
@@ -17,6 +17,7 @@ from cryptography.hazmat.decrepit.ciphers.algorithms import (
     Blowfish,
     TripleDES,
 )
+from cryptography.hazmat.decrepit.ciphers.modes import CFB, OFB
 from cryptography.hazmat.primitives import ciphers
 from cryptography.hazmat.primitives.ciphers import modes
 
@@ -137,7 +138,7 @@ class TestBlowfishModeCBC:
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        Blowfish(b"\x00" * 56), modes.OFB(b"\x00" * 8)
+        Blowfish(b"\x00" * 56), OFB(b"\x00" * 8)
     ),
     skip_message="Does not support Blowfish OFB",
 )
@@ -147,13 +148,13 @@ class TestBlowfishModeOFB:
         os.path.join("ciphers", "Blowfish"),
         ["bf-ofb.txt"],
         lambda key, **kwargs: Blowfish(binascii.unhexlify(key)),
-        lambda iv, **kwargs: modes.OFB(binascii.unhexlify(iv)),
+        lambda iv, **kwargs: OFB(binascii.unhexlify(iv)),
     )
 
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        Blowfish(b"\x00" * 56), modes.CFB(b"\x00" * 8)
+        Blowfish(b"\x00" * 56), CFB(b"\x00" * 8)
     ),
     skip_message="Does not support Blowfish CFB",
 )
@@ -163,7 +164,7 @@ class TestBlowfishModeCFB:
         os.path.join("ciphers", "Blowfish"),
         ["bf-cfb.txt"],
         lambda key, **kwargs: Blowfish(binascii.unhexlify(key)),
-        lambda iv, **kwargs: modes.CFB(binascii.unhexlify(iv)),
+        lambda iv, **kwargs: CFB(binascii.unhexlify(iv)),
     )
 
 
@@ -219,7 +220,7 @@ class TestCAST5ModeCBC:
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        CAST5(b"\x00" * 16), modes.OFB(b"\x00" * 8)
+        CAST5(b"\x00" * 16), OFB(b"\x00" * 8)
     ),
     skip_message="Does not support CAST5 OFB",
 )
@@ -229,13 +230,13 @@ class TestCAST5ModeOFB:
         os.path.join("ciphers", "CAST5"),
         ["cast5-ofb.txt"],
         lambda key, **kwargs: CAST5(binascii.unhexlify(key)),
-        lambda iv, **kwargs: modes.OFB(binascii.unhexlify(iv)),
+        lambda iv, **kwargs: OFB(binascii.unhexlify(iv)),
     )
 
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        CAST5(b"\x00" * 16), modes.CFB(b"\x00" * 8)
+        CAST5(b"\x00" * 16), CFB(b"\x00" * 8)
     ),
     skip_message="Does not support CAST5 CFB",
 )
@@ -245,7 +246,7 @@ class TestCAST5ModeCFB:
         os.path.join("ciphers", "CAST5"),
         ["cast5-cfb.txt"],
         lambda key, **kwargs: CAST5(binascii.unhexlify(key)),
-        lambda iv, **kwargs: modes.CFB(binascii.unhexlify(iv)),
+        lambda iv, **kwargs: CFB(binascii.unhexlify(iv)),
     )
 
 
@@ -297,7 +298,7 @@ class TestIDEAModeCBC:
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        IDEA(b"\x00" * 16), modes.OFB(b"\x00" * 8)
+        IDEA(b"\x00" * 16), OFB(b"\x00" * 8)
     ),
     skip_message="Does not support IDEA OFB",
 )
@@ -307,13 +308,13 @@ class TestIDEAModeOFB:
         os.path.join("ciphers", "IDEA"),
         ["idea-ofb.txt"],
         lambda key, **kwargs: IDEA(binascii.unhexlify(key)),
-        lambda iv, **kwargs: modes.OFB(binascii.unhexlify(iv)),
+        lambda iv, **kwargs: OFB(binascii.unhexlify(iv)),
     )
 
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        IDEA(b"\x00" * 16), modes.CFB(b"\x00" * 8)
+        IDEA(b"\x00" * 16), CFB(b"\x00" * 8)
     ),
     skip_message="Does not support IDEA CFB",
 )
@@ -323,7 +324,7 @@ class TestIDEAModeCFB:
         os.path.join("ciphers", "IDEA"),
         ["idea-cfb.txt"],
         lambda key, **kwargs: IDEA(binascii.unhexlify(key)),
-        lambda iv, **kwargs: modes.CFB(binascii.unhexlify(iv)),
+        lambda iv, **kwargs: CFB(binascii.unhexlify(iv)),
     )
 
 
@@ -375,7 +376,7 @@ class TestSEEDModeCBC:
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        SEED(b"\x00" * 16), modes.OFB(b"\x00" * 16)
+        SEED(b"\x00" * 16), OFB(b"\x00" * 16)
     ),
     skip_message="Does not support SEED OFB",
 )
@@ -385,13 +386,13 @@ class TestSEEDModeOFB:
         os.path.join("ciphers", "SEED"),
         ["seed-ofb.txt"],
         lambda key, **kwargs: SEED(binascii.unhexlify(key)),
-        lambda iv, **kwargs: modes.OFB(binascii.unhexlify(iv)),
+        lambda iv, **kwargs: OFB(binascii.unhexlify(iv)),
     )
 
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        SEED(b"\x00" * 16), modes.CFB(b"\x00" * 16)
+        SEED(b"\x00" * 16), CFB(b"\x00" * 16)
     ),
     skip_message="Does not support SEED CFB",
 )
@@ -401,5 +402,5 @@ class TestSEEDModeCFB:
         os.path.join("ciphers", "SEED"),
         ["seed-cfb.txt"],
         lambda key, **kwargs: SEED(binascii.unhexlify(key)),
-        lambda iv, **kwargs: modes.CFB(binascii.unhexlify(iv)),
+        lambda iv, **kwargs: CFB(binascii.unhexlify(iv)),
     )