+++ tests/hazmat/primitives/test_ciphers.py
@@ -11,6 +11,7 @@ import pytest
 
 from cryptography import utils
 from cryptography.exceptions import AlreadyFinalized
+from cryptography.hazmat.decrepit.ciphers.modes import CFB, CFB8, OFB
 from cryptography.hazmat.primitives import ciphers
 from cryptography.hazmat.primitives.ciphers import modes
 from cryptography.hazmat.primitives.ciphers.algorithms import (
@@ -52,9 +53,7 @@ class TestAES:
 
 
 class TestAESXTS:
-    @pytest.mark.parametrize(
-        "mode", (modes.CBC, modes.CTR, modes.CFB, modes.CFB8, modes.OFB)
-    )
+    @pytest.mark.parametrize("mode", (modes.CBC, modes.CTR, CFB, CFB8, OFB))
     def test_invalid_key_size_with_mode(self, mode, backend):
         with pytest.raises(ValueError):
             ciphers.Cipher(AES(b"0" * 64), mode(b"0" * 16), backend)